+++ components/server/Cargo.toml
@@ -57,7 +57,6 @@ grpcio-health = { version = "0.10", default-features = false, features = ["proto
 hex = "0.4"
 keys = { path = "../keys", default-features = false }
 kvproto = { git = "https://github.com/pingcap/kvproto.git" }
-lazy_static = "1.3"
 libc = "0.2"
 log = { version = "0.4", features = ["max_level_trace", "release_max_level_debug"] }
 log_wrappers = { path = "../log_wrappers" }
@@ -71,7 +70,7 @@ rand = "0.8"
 resolved_ts = { path = "../../components/resolved_ts", default-features = false }
 resource_metering = { path = "../resource_metering" }
 security = { path = "../security", default-features = false, features = ["tonic"] }
-serde_json = { version = "1.0", features = ["preserve_order"] }
+serde_json = "1.0"
 slog = { version = "2.3", features = ["max_level_trace", "release_max_level_debug"] }
 slog-global = { version = "0.1", git = "https://github.com/breeswish/slog-global.git", rev = "d592f88e4dbba5eb439998463054f1a44fbf17b9" }
 tempfile = "3.0"
